Infraestructura completa: EF Core, repositorios migraciones y seed admin

+++ WebApi-HU3.Infraestructure/Repositories/StudentRepository.cs
@@ -1,41 +1,59 @@
+using Microsoft.EntityFrameworkCore;
+using WebApi_HU3.Domain.Entities;
+using WebApi_HU3.Domain.Interfaces;
 using WebApi_HU3.Infraestructure.Data;
 
 namespace WebApi_HU3.Infraestructure.Repositories;
 
+//inyeccion de contexto base de datos 'DbContext'
 public class StudentRepository : IStudentRepository
 {
     private readonly AppDbContext _context;
-
+    
+    
+    //se obtiene lista completa de los estudiantes registrados
     public StudentRepository(AppDbContext context)
     {
         _context = context;
     }
 
-    public async Task<Student?> GetByIdStudentAsync(int id)
+    //se obtiene lista completa de los estudiantes registrados
+    public async Task<IEnumerable<Student>> GetAllStudents()
     {
-        await _context.Students.FindAsync(id);
+        return await _context.Students.ToListAsync();
     }
-
-    public async Task<IEnumerable<Student>> GetAllStudentAsync()
+    
+    //se obtiene un estudiante por id
+    public async Task<Student?> GetStudentById(int id)
     {
-        await _context.Students.ToListAsync();
+        return await _context.Students.FindAsync(id);
     }
-
-    public async Task AddStudentAsync(Student student)
+    
+    //se agregar un nuevo estudiante
+    public async Task<Student> AddStudent(Student student)
     {
         await _context.Students.AddAsync(student);
         await _context.SaveChangesAsync();
+        return student;
     }
 
-    public async Task UpdateStudentAsync(Student student)
+    //actualiza informacion de un estudiante existente
+    public async Task<Student> UpdateStudent(Student student)
     {
         _context.Students.Update(student);
         await _context.SaveChangesAsync();
+        return student;
     }
 
-    public async Task DeleteStudentAsync(Student student)
+    //elimina un estudiante por id
+    public async Task<Student> DeleteStudent(int id)
     {
-        _context.Students.Remove(student);
-        await _context.SaveChangesAsync();
+        var student = await _context.Students.FindAsync(id);
+        if (student != null)
+        {
+            _context.Students.Remove(student);
+            await _context.SaveChangesAsync();
+        }
+        return student;
     }
 }
